Refactor View interface out of the lens module
Change-type: patch
